JSON

+++ Json/a.h
@@ -13,7 +13,7 @@ public:
     void setNome(const QString nome);
     QString getNome(const int index) const;
     qint8 getId() const;
-    void setId(const qint8 id);
+    void setId(const qint16 id);
     void addItem(const int item);
     int getItem(const int index) const;
     void read(const QJsonObject &json);
@@ -21,7 +21,7 @@ public:
 
 private:
     QString nome;
-    qint8 id;
+    qint16 id;
     QVector<int> tabela;
 };
